Create TextCloud.py
This file will crawl through several urls of given depth and return a text cloud of the most prevalent words in the webpages. The words are filtered to not include punctuation, and to combine words of the same root without various endings. It also removes common words ("a", "the", "and"). The program then sets up a window that includes the text cloud with word prevalence indicated by size.